Add files via upload
For this project, we were given data on molecular expression during embryonic development. The images that were given to us had labels associated with them dictating whether the image was a good or a bad signal. Our goal for this project was to build a machine that was capable of discerning which images were useful signals.